requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
colorama==0.4.6
Django==4.1.2
django-bootstrap-v5==1.0.11
django-phone-field==1.8.1
Pillow==9.2.0
qrcode==7.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.5
